fix: Partially revert back to double for password
Revert back to double for the password on the home page and on mobx store,
cause an error on compiling due to the password slider for the length.

+++ lib/pages/home_page.dart
@@ -215,7 +215,7 @@ class _HomePageState extends State<HomePage> {
                           onChanged: (double value) {
                             setState(() {
                               passwordStore
-                                ..changeLenght(value.toInt())
+                                ..changeLenght(value)
                                 ..generateNew();
                             });
                           })),
